microservices/predict-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
PyJWT==2.8.0